feat: add KSPP compliance check script
We tailor results for our needs. I picked Python for the script as we
don't have jq in the tools ;)

Also fixed the kernel config to be compliant according to the script and
subset of the enforcements.

I had to add a patch to kernel's Kconfig files to switch off some
default options which otherwise get enabled via `olddefconfig` and
violate KSPP.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>