feat: add support for ODroid HC4
Adds the kernel module required to ensure proper functioning of the
ODroid HC4's GPU so that HDMI output works as expected.

Adds the kernel module required to ensure proper functioning of the SATA
ODroid HC4's SATA ports.

Adds a u-boot variant for the ODroid HC4. This is not used directly, but
rather flashed into the SPI of the unit so that it can boot Talos.
Documentation on this process can be found in the Talos repo.

Signed-off-by: Brandon McNama <brandonmcnama@outlook.com>
Signed-off-by: Noel Georgi <git@frezbo.dev>