chore: copy over the kernel signing public key
Copy over the kernel signing public key so that it can be used to verify
modules are signed by the correct key. The signing key is copied over as
otherwise in-order to extract the signing key we either need the
uncompressed `vmlinux` (which is super huge to copy over) or need the
`System.map` file along with `vmlinuz` and a perl script to extract it.
It's just easier to just copy over the `signing_key.x509` file.

Signed-off-by: Noel Georgi <git@frezbo.dev>