feat: enable modules for mlx infiniband
mlx4 and mlx5 core and ethernet modules are enabled.
Infiniband and IPoIB are also enabled.
And this patch enables the remaining bits to get most Infiniband
hardware working out of the box.

Requires inclusion of the new modules in the list for initramfs to
support cases like booting from or otherwise early use of
the Infiniband network.

Signed-off-by: Radosław Piliszek <radoslaw.piliszek@gmail.com>
Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>