feat: Add kernel-devel package for building external kernel modules
This recipe is based on the Fedora kernel-devel RPM build spec [1].

The purpose of this image is to have a published smaller image to use as
a prerequisite when building external modules for the corresponding
kernel.

As a comparison, the size of this kernel-devel image for amd64 is 96 MiB
vs 12.8 GiB for the complete kernel-build intermediate image with all of
the sources and object files.

[1]: https://src.fedoraproject.org/rpms/kernel/raw/rawhide/f/kernel.spec